Moved trie spec into its own file
